zaki

+++ dda.c
@@ -6,7 +6,7 @@
 /*   By: momihamm <momihamm@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/06 19:01:43 by momihamm          #+#    #+#             */
-/*   Updated: 2024/03/02 02:12:59 by momihamm         ###   ########.fr       */
+/*   Updated: 2024/03/11 04:29:59 by momihamm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,6 +40,14 @@ void	dda(t_ray *obj, double x1, double y1)
 	int	indx;
 
 	indx = 0;
+	// if (x1 < 0)
+	// 	x1 = 0;
+	// if (y1 < 0)
+	// 	y1 = 0;
+	// if (x1 > 1440)
+	// 	x1 = 1440;
+	// if (y1 > (obj->the_rows * 32))
+	// 	y1 = (obj->the_rows * 32);
 	obj->algo->x_zero = (obj->plays->x_play) + 16;
 	obj->algo->y_zero = (obj->plays->y_play) + 16;
 	obj->algo->dx = x1 - obj->algo->x_zero;
@@ -48,8 +56,8 @@ void	dda(t_ray *obj, double x1, double y1)
 		obj->algo->steps = fabs(obj->algo->dx);
 	else
 		obj->algo->steps = fabs(obj->algo->dy);
-	obj->algo->x_incre = obj->algo->dx / (double) obj->algo->steps;
-	obj->algo->y_incre = obj->algo->dy / (double) obj->algo->steps;
+	obj->algo->x_incre = obj->algo->dx / obj->algo->steps;
+	obj->algo->y_incre = obj->algo->dy / obj->algo->steps;
 	obj->algo->flo_x = obj->algo->x_zero;
 	obj->algo->flo_y = obj->algo->y_zero;
 	while (indx <= obj->algo->steps)
